gnome-extra/gtkhtml: 4.5.5 → 4.6.0 and add live ebuild

+++ gnome-extra/gtkhtml/gtkhtml-4.6.0.ebuild
@@ -1,19 +1,24 @@
 # Copyright 1999-2012 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/gnome-extra/gtkhtml/gtkhtml-4.4.3.ebuild,v 1.1 2012/06/23 09:46:13 pacho Exp $
+# $Header: $
 
 EAPI="4"
 GCONF_DEBUG="no"
 GNOME2_LA_PUNT="yes"
 
-inherit gnome2 eutils
+inherit gnome2
+[[ ${PV} = 9999 ]] && inherit gnome2-live
 
 DESCRIPTION="Lightweight HTML Rendering/Printing/Editing Engine"
 HOMEPAGE="http://www.gnome.org/"
 
-LICENSE="GPL-2 LGPL-2"
+LICENSE="GPL-2+ LGPL-2+"
 SLOT="4.0"
-KEYWORDS="~alpha ~amd64 ~arm ~ia64 ~ppc ~ppc64 ~sh ~sparc ~x86 ~x86-fbsd ~x86-freebsd ~amd64-linux ~ia64-linux ~x86-linux"
+if [[ ${PV} = 9999 ]]; then
+	KEYWORDS=""
+else
+	KEYWORDS="~alpha ~amd64 ~arm ~ia64 ~ppc ~ppc64 ~sh ~sparc ~x86 ~x86-fbsd ~x86-freebsd ~amd64-linux ~ia64-linux ~x86-linux"
+fi
 IUSE=""
 
 # orbit is referenced in configure, but is not used anywhere else
@@ -40,8 +45,10 @@ pkg_setup() {
 
 src_prepare() {
 	# Regenerate marshallers for <glib-2.31 compatibility
-	rm -v components/editor/gtkhtml-spell-marshal.{c,h} \
-		components/editor/gtkhtml-editor-marshal.{c,h} || die
+	if [[ ${PV} != 9999 ]]; then
+		rm -v components/editor/gtkhtml-spell-marshal.{c,h} \
+			components/editor/gtkhtml-editor-marshal.{c,h} || die
+	fi
 	gnome2_src_prepare
 }
